Fix the bug 1333238: OpenStack console shows some instances in 'migrating' state forever when migration fails for some reason
After analyzing the source code, I only could move one part of the code into pre_live_migration.
But the remaining source code can not be moved,
because the number of the parameters in the method 'check_can_live_migration ' and the method 'pre_live_migration' are different.
Some necessary parameters can not be passed by the nova framework.

Closes-Bug: 1333238
Change-Id: I6cc59d76dca2eab5e22669fd0d745cb8bda53c62