Timing issue in remote-keystone deployment model.
This is actually timing issue between openstack service endpoints becoming
active and powervc-driver's client initialization of those.

1. get_client() call with version_filter might runs into the 'None' client
error mentioned in the bug, which means *currently* only glance sync is
affected. When no version filtered, we do a rediscover for the svc_type.

2. get_client() call without version_filter or new_client() call just choose
the latest avaiable version service, so no 'None' client error, although the
chosen version might not be the real *latest* but the default hardcoded 'v1'
version, due to the timing issue mentioned in the bug. For the latter case,
unless there is an observation/notification mechanism or poll mechanism for
service versions update, based on current design, there is no perfect solution
as far as I can imagine.

Change-Id: I9ca0f906c8954ecd621be22445c12fd6a55a1ae7
Closes-Bug: 1358215