Temp fix for powervc side volumes operation not synced to local Openstack in cinder-powervc
Currently cinder service uses threading instead of eventlet's event,
This causes if you don't monkey patch threading, it will hang there
all the time. For detail, check this fix in incubator project:
https://github.com/openstack/oslo-incubator/commit/\
1e7007824374842bc8108dff5e5e2694e12932b2
Will create a bug to cinder to fix this. Temporarily fix is to replace
cinder service instead of nova service in 1365373.

Change-Id: Ibe770e51563f43495228d84dd6d0284a62bba9e8
Closes-Bug: #1365373