Add/Remove/Modify vNIC post deployment Edit
User call the following hosting cli to add a vNIC with IP and port
specified:
    nova interface-attach --net-id {net_uuid} --fixed-ip {ipAddress}
<server_name>
PowerVC driver handle the request in the following driver method:
    def attach_interface(self, instance, image_meta, network_info):
then call the PowerVC restAPI to attach the interface to server.

User call the following hosting cli to remove a vNIC with port
specified:
    nova interface-detach <server-id> <port-id>
PowerVC driver handle the request in the following driver method:
    def detach_interface(self, instance, network_info):
then call the PowerVC restAPI to attach the interface to server.

Change-Id: If0ae14a86005da06dc35620c2af47f37112228d5
Closes-Bug: #1375132