Replace glance V1 API calls to use V2 API
With the Queens release of Openstack, the community removed the
support for v1 glance API. Thus, we need to fix the corresponding
implementation in the PowerVC Driver using glance v1 calls to v2 calls.
As part of these changes, we've removed all the v1 glance client calls
to v2, and also observed that the response format has changed as well
and these changes should address all of that.

Change-Id: Ic76bad59ba3da7178acaeeb093113358e50c6169
Closes-Bug: #1783096