Custom properties are missing in the image details section.
This is due to the changes made to handle the
community change where V1 glance API got removed
with Queens release of openstack. And, with V2
the response format has changed returning fewer
properties.This bug is to fix the issue related
to missing  custom properties and add them back
to be displayed to the user.

Change-Id: I116ad05b00de6e94cf6c507dbcffa607b916fc42
Closes-Bug: #1814739